+++ changelog.md
@@ -4,8 +4,9 @@
 - Added category filter to the Tag list. New tags automatically get assigned the current filter as category name.
 - Added higlight blend modes "Tint", "Tint Bright Colors" and "Dither"
 - Some blend modes transparency can now be set by the selected fill color's alpha
-- Fixed color picker showing solid colors with transparency
 - Added text rendering anti alias options.
+- Fixed color picker showing solid colors with transparency
+- Fixed "out of memory" error when selecting a region that extended off screen
 
 ## 2.5
 - Trim the number of thumbnails in the list periodically, based on max thumbs setting